+++ Scripts/Components/DefenceComponent.cs
@@ -4,13 +4,13 @@ using System;
 [Serializable]
 public struct DefenceComponent
 {
-    public int hp;
-    public int maxHP;
-    public int manaShield;
-    public int maxManaShield;
-    public int armor;
-    public int hpRegen;
-    public int manaShieldRegen;
+    public float hp;
+    public float maxHP;
+    public float manaShield;
+    public float maxManaShield;
+    public float armor;
+    public float hpRegen;
+    public float manaShieldRegen;
     public float timeToStartHpRegenAfterTakeDamage;
     public float timeToStartManaShieldRegenAfterTakeDamage;
     public ArmorTypeEnum armorType;
